fix: Add missing assembly refs, fix GameState comparison, and Unity 6 API updates
- Added Unity.Collections to Laboratory.Editor (required by Unity.Entities)
- Added Laboratory.Core.Management to Editor (for GameManager)
- Added Laboratory.Gameplay to Editor (for Pickup)
- Fixed GameState comparison using (int) cast (two different enums with same name)
- Fixed ulong to long conversion in BuildOptimizationTool
- Disabled platformKeyword API in ShaderVariantStripper (obsolete in Unity 6)

+++ Assets/_Project/Scripts/Editor/Tools/BuildOptimizationTool.cs
@@ -585,7 +585,7 @@ namespace Laboratory.Editor.Tools
             if (_generateReport && _lastReport != null)
             {
                 _lastReport.buildTime = (float)(System.DateTime.Now - _lastReport.buildStartTime).TotalSeconds;
-                _lastReport.totalSize = report.summary.totalSize;
+                _lastReport.totalSize = (long)report.summary.totalSize;
 
                 foreach (var file in report.files)
                 {
